requirements.txt
fastapi==0.115.12
httpx==0.28.1
pip-chill==1.0.3
pymongo==4.12.0
pyspark==3.5.5
python-dotenv==1.1.0
requests==2.32.3
uvicorn==0.34.2